test: test scraper,data processing , prediction and function modules
