+++ pyproject.toml
@@ -16,9 +16,12 @@ facebook-scraper = "^0.2.49"
 jupyterlab = "^3.2.2"
 ipynb = "^0.5.1"
 import-ipynb = "^0.1.3"
+coverage = "^6.1.1"
+pytest-cov = "^3.0.0"
 
 [tool.poetry.dev-dependencies]
 pytest = "^5.2"
+autopep8 = "^1.6.0"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
