+++ socialreact/analyzer/analyzerr.py
@@ -1,0 +1,16 @@
+# class will handle data processing including 
+# *labeling 
+# *feature extraction
+# *split features to train and test datasets using SKLEARN
+# *calculate Accuracy using train and test data usng support vector machine or random forest
+# !============================================
+# ? Neural Network
+# * create a Sequential model 
+# * data preparation
+# * compile the model
+# * Train The Model
+# * Testing the model
+# * retrieveing score and accuracy
+# * validate model accuracy positive, negative score
+# * create a funtion to predict given post positive - negative impact
+
